+++ ko_working_app/nuxt.config.ts
@@ -2,5 +2,10 @@
 export default defineNuxtConfig({
   compatibilityDate: '2024-11-01',
   devtools: { enabled: true },
-  modules: ['@nuxt/ui']
+  modules: ['@nuxt/ui'],
+  runtimeConfig: {
+    public: {
+      hostUrl: "http://localhost:8080"
+    }
+  }
 })
